fix(Tewodros.imp-blog-rate): fix merge conflict

+++ Backend/starter_project_zsilentcoders/BlogApp3/BlogApp.Application/Features/BlogRates/CQRS/Handlers/DeleteBlogRateCommandHandler.cs
@@ -1,33 +1,62 @@
-﻿using AutoMapper;
+﻿using System.ComponentModel.DataAnnotations;
 using BlogApp.Application.Contracts.Persistence;
-using BlogApp.Application.Features.BlogRates.CQRS.Commands;
+using BlogApp.Application.Features.Blog.DTOs.Validators;
+using BlogApp.Application.Features.Blog.CQRS.Requests.Commands;
+using AutoMapper;
+using BlogApp.Application.Exceptions;
+using BlogApp.Application.Responses;
 using MediatR;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+using BlogApp.Application.Features.BlogRates.CQRS.Commands;
+using BlogApp.Application.Features.BlogRates.DTOs.Validators;
 
-namespace BlogApp.Application.Features.BlogRates.CQRS.Handlers
+namespace Application.Features.BlogRates.Handlers.Commands;
+
+public class DeleteBlogRateCommandHandler : IRequestHandler<DeleteBlogRateCommand, Result<Unit>>
 {
-    public class DeleteBlogRateCommandHandler : IRequestHandler<DeleteBlogRateCommand, Unit>
-    {
-        private readonly IUnitOfWork _unitOfWork;
-        private readonly IMapper _mapper;
-        public DeleteBlogRateCommandHandler(IUnitOfWork unitOfWork, IMapper mapper)
-        {
-            _unitOfWork = unitOfWork;
-            _mapper = mapper;
-            
-        }
-        public async Task<Unit> Handle(DeleteBlogRateCommand request, CancellationToken cancellationToken)
-        {
-            var blogRate = await _unitOfWork.BlogRateRepository.Get(request.Id);
-            await _unitOfWork.BlogRateRepository.Delete(blogRate);
 
-            return Unit.Value;
-                
-            
+    private readonly IUnitOfWork _unitOfWork;
+    private readonly IMapper _mapper;
+
+    public DeleteBlogRateCommandHandler(IUnitOfWork unitOfWork, IMapper mapper)
+    {
+        _unitOfWork = unitOfWork;
+        _mapper = mapper;
+    }
+
+    public async Task<Result<Unit>> Handle(DeleteBlogRateCommand request, CancellationToken cancellationToken)
+    {
+        var response = new Result<Unit>();
+
+        var validator = new DeleteBlogRateDtoValidator(_unitOfWork);
+        var validationResult = await validator.ValidateAsync(request.DeleteBlogRateDto);
+
+        if (validationResult.IsValid == false)
+        {
+            response.Success = false;
+            response.Message = "Deletion Failed";
+            response.Errors = validationResult.Errors.Select(q => q.ErrorMessage).ToList();
         }
+        else
+        {
+            
+
+            var blogRate = await _unitOfWork.BlogRateRepository.GetBlogRateByBlogAndRater(request.DeleteBlogRateDto.BlogId, request.DeleteBlogRateDto.RaterId);
+            await _unitOfWork.BlogRateRepository.Delete(blogRate);
+            if (await _unitOfWork.Save() > 0)
+            {
+                response.Message = "Deletion Successful!";
+                response.Value = new Unit();
+
+            }
+            else
+            {
+                response.Success = false;
+                response.Message = "Deletion Failed";
+                response.Errors = validationResult.Errors.Select(q => q.ErrorMessage).ToList();
+
+            }
+        }
+
+        return response;
     }
 }
